+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Fixed multiple value tag filtering. [#346](https://github.com/Microsoft/PSRule/issues/346)
+- Added filtering for rules against a baseline with `Get-PSRule`. [#345](https://github.com/Microsoft/PSRule/issues/345)
+
 ## v0.12.0-B1912002 (pre-release)
 
 - Fixed TargetType fall back to type name. [#339](https://github.com/Microsoft/PSRule/issues/339)